% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_plotdim.R
\name{anchor}
\alias{anchor}
\title{Add regex line-start and -end.}
\usage{
anchor(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A character vector.
}
\description{
Add regex line-start and -end.
}
\examples{
anchor(c("a", "b"))
}
\seealso{
Other general functions to edit data in place: \code{\link{fill_na}},
  \code{\link{glue_comma}}, \code{\link{suffix_match}},
  \code{\link{to_tidy_names}}
}
\concept{general functions to edit data in place}
\keyword{internal}
