% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crucial_names.R
\name{check_crucial_names}
\alias{check_crucial_names}
\title{Check if an object contains specific names.}
\usage{
check_crucial_names(x, nms)
}
\arguments{
\item{x}{A named object.}

\item{nms}{String; names expected to be found in \code{x}.}
}
\value{
Invisible \code{x}, or an error with informative message.
}
\description{
Check if an object contains specific names.
}
\examples{
v <- c(x = 1)
check_crucial_names(v, "x")

dfm <- data.frame(x = 1)
check_crucial_names(dfm, "x")
}
\seealso{
Other predicates: \code{\link{detect_if}},
  \code{\link{has_table_names}}, \code{\link{is_multiple}}

Other functions to check inputs.: \code{\link{is_multiple}}

Other functions for developers.: \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{insensitive}}, \code{\link{is_multiple}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}, \code{\link{round_any}}
}
\concept{functions for developers.}
\concept{functions to check inputs.}
\concept{predicates}
