% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_censusid.R
\name{collapse_censusid}
\alias{collapse_censusid}
\title{Collapse the values of CensusID.}
\usage{
collapse_censusid(x)
}
\arguments{
\item{x}{A dataframe -- specifically a ViewFullTable.}
}
\value{
A modified version of \code{x}, most likely with less rows.
}
\description{
This function is useful to summarize the status history of a tree tag across
censuses. It is particularly useful to avoid duplicated tree tags on maps.
}
\examples{
x <- data.frame(
  CensusID = c(1, 1, 1, 1, 2, 2, 2, 2),
  Tag = c(1, 1, 2, 2, 1, 1, 2, 2),
  Status = c("alive", "dead", "dead", "dead", "alive", "alive", "alive", "dead"),
  stringsAsFactors = FALSE
)
x
collapse_censusid(x)
}
