% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn_na.R
\name{condition_na}
\alias{condition_na}
\alias{warn_na}
\alias{inform_na}
\alias{abort_na}
\title{Throw condition if variable has missing values.}
\usage{
warn_na(x)

inform_na(x)

abort_na(x)
}
\arguments{
\item{x}{A vector or dataframe.}
}
\value{
A warning message and invisible \code{x}.
}
\description{
Throw condition if variable has missing values.
}
\examples{
warn_na(c(x = 1, y = NA))
inform_na(data.frame(x = 1, y = NA))
}
\seealso{
Other functions to throw conditions: \code{\link{flag_if}}
}
\concept{functions to throw conditions}
