% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_multiple.R
\name{detect_multiple}
\alias{detect_multiple}
\alias{flag_multiple}
\title{Detect and flag multiple values of a variable.}
\usage{
detect_multiple(.data)

flag_multiple(.data, cond, msg = NULL)
}
\arguments{
\item{.data}{A dataframe.}

\item{cond}{Symbol; the bare name of a function that outputs a condition:
e.g. warning, stop, message, rlang::warn, rlang::abort, rlang::inform.}

\item{msg}{String; a custom message.}

\item{name}{String; the name of a variable of \code{.data}.}
}
\value{
\itemize{
\item \code{detect_multiple()}:A single \code{TRUE} or \code{FALSE}. Insensitive to upper-
lower-case.
}
}
\description{
\itemize{
\item \code{detect_multiple()} is a predicate function that returns \code{TRUE} if it
detects multiple different values of a variable.
\item \code{flag_multiple()} throws a condition and, invisibly, the input data.
}
}
\examples{
# DETECT -----------------------------------------------------------------
detect_multiple(c(1, 2))
detect_multiple(c(1, 1))
detect_multiple(c(1, NA))

# FLAG -------------------------------------------------------------------
# On a vector
# Silent
single_value <- c(1, 1, 1)
flag_multiple(single_value, warning)

num <- c(1:3)
flag_multiple(num, warning)
chr <- c(letters[1:3])
flag_multiple(chr, message, "Hello world.")
}
\seealso{
\code{\link[=detect_multiple_f]{detect_multiple_f()}}, \code{\link[=flag_multiple_f]{flag_multiple_f()}}.

Other functions to check inputs.: \code{\link{check_crucial_names}}

Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{detect_multiple_f}},
  \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{insensitive}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}, \code{\link{round_any}}

Other predicates.: \code{\link{extract_insensitive}}
}
\concept{functions for developers.}
\concept{functions to check inputs.}
\concept{predicates.}
