% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_strings.R
\name{enline}
\alias{enline}
\title{Add regex line-start and -end.}
\usage{
enline(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A character vector.
}
\description{
Add regex line-start and -end.
}
\examples{
enline(c("a", "b"))
}
