% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_path_factory.R
\name{example_path_factory}
\alias{example_path_factory}
\title{Create functions to output paths to directories from inst/extdata.}
\usage{
example_path_factory(package, .f)
}
\arguments{
\item{package}{Character string giving the package name where to look for the
desired data.}

\item{.f}{A function to apply to a path. Tested functions include \code{file.path}
, \code{basename} and \code{dirname}.}
}
\value{
A string giving the \code{path} from inst/extdata/ to a file with
extension, even if the output of the closure will not be the file itself but
the containing directory.
}
\description{
This is a function factory that creates other functions. It is intended not
for users but for developers. The goal is to quickly and consistently
create functions that output paths to directories from inst/data -- where
the system stores example data.
}
\examples{
# Similar to fs::path
example_path <- example_path_factory("fgeo.base", file.path)
example_path(path = "folder/file.csv")

# Similar to fs::path_file
example_path_file <- example_path_factory("fgeo.base", basename)
example_path_file(path = "folder/file.csv")

# Similar to fs::path_dir
example_path_dir <- example_path_factory("fgeo.base", dirname)
example_path_dir(path = "folder/file.csv")
}
