% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms_restore_matching.R
\name{extract_insensitive}
\alias{extract_insensitive}
\alias{detect_insensitive}
\title{Detect and extract matching strings ignoring case.}
\usage{
extract_insensitive(x, ref)

detect_insensitive(x, ref)
}
\arguments{
\item{x}{A string to be muted as in \code{ref}, it a case insensitive match is
found.}

\item{ref}{A string to use as a reference to match \code{x}.}
}
\value{
\code{detect_*} and \code{extract_*} return a logical vector and a string.
}
\description{
Detect and extract matching strings ignoring case.
}
\examples{
new <- c("stemid", "n")
reference <- c("StemID", "treeID")
detect_insensitive(new, reference)
extract_insensitive(new, reference)

vft <- data.frame(TreeID = 1, Status = 1)
extract_insensitive(tolower(names(vft)), names(vft))
extract_insensitive(names(vft), tolower(names(vft)))
}
\seealso{
Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{check_unique}}, \code{\link{exists_in_pkg}},
  \code{\link{insensitive}},
  \code{\link{nms_restore_matching}},
  \code{\link{round_any}}
}
\concept{functions for developers.}
