% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.base-package.R
\docType{package}
\name{fgeo.base-package}
\alias{fgeo.base}
\alias{fgeo.base-package}
\title{fgeo.base: ForestGEO Functions With No External Dependency}
\description{
ForestGEO functions with no external dependency.
    These functions are isolated from other packages so they can be
    imported by any fgeo package without dragging any external dependency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.base}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.base/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
