% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_datasets.R
\name{find_datasets}
\alias{find_datasets}
\title{Find datasets in a package.}
\usage{
find_datasets(package)
}
\arguments{
\item{package}{String; package name to search.}
}
\value{
Sting.
}
\description{
Find datasets in a package.
}
\examples{
find_datasets("fgeo.base")
}
\concept{general functions to find or approximate}
