% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_multiple.R
\name{flag_multiple}
\alias{flag_multiple}
\alias{flag_multiple_vector}
\title{Report if a vector or a variable of a dataframe has muliple distinct values.}
\usage{
flag_multiple(x, x_var, cond = warning, msg = NULL)

flag_multiple_vector(v, cond, msg = NULL)
}
\arguments{
\item{x}{A dataframe.}

\item{x_var}{String; the name of a variable of \code{x}.}

\item{cond}{String; the name of a function that outputs a condition: one of
"warning", "stop", "message".}

\item{msg}{String; a custom message.}

\item{v}{A vector.}
}
\value{
Invisible \code{v} or a condition and a message.
}
\description{
Report if a vector or a variable of a dataframe has muliple distinct values.
}
\examples{
# On a vector
unique_v <- rep(1, 3)
num <- c(1:3)
chr <- c(letters[1:3])
flag_multiple_vector(unique_v, warning)
flag_multiple_vector(num, warning)
flag_multiple_vector(chr, message, "Hello world.")

# On a dataframe
.df <- data.frame(a = 1:3, b = 1, stringsAsFactors = FALSE)

flag_multiple(.df, "a")
flag_multiple(.df, "a", message, "Hello world.")
# Silent
flag_multiple(.df, "b", warning, "Hello world")

\dontrun{
# Dealing with grouped data
if (!requireNamespace("dplyr")) {
  # `b` is single within groups but multiple accross entire dataset
  .df <- data.frame(
    a = c(1, 1, 2, 2), b = c(1, 1, 2, 2),
    stringsAsFactors = FALSE
  )

  by_x <- dplyr::group_by(.df, a)
  # Works accross entire dataset
  flag_multiple(by_x, "b")
  # Works within groups entire dataset
  dplyr::do(by_x, (flag_multiple(., "b")))
  # Also consider tidyr::nest() + dplyr::mutate() + dpyr::map())
}
}
}
\seealso{
Other functions to check inputs.: \code{\link{check_crucial_names}}

Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{insensitive}}, \code{\link{multiple_var}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}, \code{\link{round_any}}
}
\concept{functions for developers.}
\concept{functions to check inputs.}
