% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_multiple.R
\name{flag_multiple_f}
\alias{flag_multiple_f}
\title{Flag if a vector or a variable of a dataframe has multiple distinct values.}
\usage{
flag_multiple_f(name, cond = warning)
}
\arguments{
\item{name}{String; the name of a variable of \code{.data}.}

\item{cond}{Symbol; the bare name of a function that outputs a condition:
e.g. warning, stop, message, rlang::warn, rlang::abort, rlang::inform.}

\item{.data}{A dataframe.}

\item{msg}{String; a custom message.}
}
\value{
Invisible.
}
\description{
Flag if a vector or a variable of a dataframe has multiple distinct values.
}
\examples{
# On a vector
# Silent
single_value <- c(1, 1, 1)
flag_multiple(single_value, warning)

num <- c(1:3)
flag_multiple(num, warning)
chr <- c(letters[1:3])
flag_multiple(chr, message, "Hello world.")

# On a dataframe
.df <- data.frame(a = 1:3, b = 1, stringsAsFactors = FALSE)

flag_multiple_f("a")(.df)
flag_multiple_f("a", message)(.df, "Hello world.")
# Silent
flag_multiple_f("b", warning)(.df, "Hello world")

\dontrun{
# Dealing with grouped data
if (!requireNamespace("dplyr")) {
  library(dplyr)
  # `b` is single within groups but multiple accross entire dataset
  .df <- tibble(a = c(1, 1, 2, 2), b = c(1, 1, 2, 2))

  by_x <- group_by(.df, a)
  # Works accross entire dataset -- expect warning
  flag_if_multiple_b <- flag_multiple_f("b")
  flag_if_multiple_b(by_x)
  # Works within groups -- expect silent
  do(by_x, (flag_if_multiple_b(.)))
  # Same, with interface similar to lapply() and purrr::map()
  fgeo.tool::by_group(by_x, flag_if_multiple_b)
  # Also consider tidyr::nest() + dplyr::mutate() + dpyr::map())
}
}
}
\seealso{
Other functions to check inputs.: \code{\link{check_crucial_names}}

Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{detect_multiple_f}},
  \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{insensitive}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}, \code{\link{round_any}}
}
\concept{functions for developers.}
\concept{functions to check inputs.}
