% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_matrix.R
\name{gather_matrix}
\alias{gather_matrix}
\alias{gather_matrix_ls}
\title{Gather matrix columns into a long-format dataframe.}
\usage{
gather_matrix(...)

gather_matrix_ls(mat, rownm = "rownames", colnm = "colnames",
  value = "value")
}
\arguments{
\item{...}{Arguments passed to gather_matrix_ls.}

\item{mat}{A matrix.}

\item{rownm, colnm, value}{Names of the columns that store the row names, the
column names and the values.}
}
\value{
\itemize{
\item \code{\link[=gather_matrix]{gather_matrix()}}: A dataframe.
\item \code{\link[=gather_matrix_ls]{gather_matrix_ls()}}: A list of dataframes.
}
}
\description{
These functions convert a matrix to dataframes with three-columns: One for
the matrix-rownames; a second one for the matrix-colnames; and a third one
for the matrix-values. These functions combine features of \code{tidyr::gather()}
and \code{tibble::enframe()}:
\itemize{
\item \code{\link[=gather_matrix]{gather_matrix()}} outputs a long-format dataframe.
\item \code{\link[=gather_matrix_ls]{gather_matrix_ls()}}
returns a list of dataframes where each element of the list maps to a column
of the original matrix.
}
}
\examples{
 mat <- matrix(1:6, 2, dimnames = list(LETTERS[1:2], letters[1:3]))
 mat

 gather_matrix_ls(mat)

 gather_matrix(mat)

 gather_matrix(mat, "metric", "sp")

 mat <- matrix(1:6, 2)
 mat

 gather_matrix(mat)
}
\seealso{
\code{enframe()} (\strong{tibble} package) and \code{gather()} (\strong{tidyr} package).
}
