% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_comma.R
\name{glue_comma}
\alias{glue_comma}
\alias{glue_pipe}
\title{Collapse strings into a single one.}
\usage{
glue_comma(...)

glue_pipe(...)
}
\arguments{
\item{...}{Arguments passed on to \code{base::paste}
\describe{
  \item{sep}{a character string to separate the terms.  Not
    \code{\link{NA_character_}}.}
  \item{collapse}{an optional character string to separate the results.  Not
    \code{\link{NA_character_}}.}
}}
}
\value{
String.
}
\description{
These functions are a shortcut for paste0(..., collapse = "<somethig>").
}
\examples{
glue_comma(1:3)
glue_pipe(1:3)
}
