% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_plotdim.R
\name{guess_max}
\alias{guess_max}
\title{Guess maximum value of a vector with flexible accuracy.}
\usage{
guess_max(x, accuracy)
}
\arguments{
\item{x}{Numeric vector.}

\item{accuracy}{A single number.}
}
\value{
A number.
}
\description{
Guess maximum value of a vector with flexible accuracy.
}
\examples{
guess_max(1:19, 20)
}
\seealso{
Other general functions to find or approximate: \code{\link{find_datasets}},
  \code{\link{guess_plotdim}}, \code{\link{round_any}}
}
\concept{general functions to find or approximate}
