% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_table_names.R
\name{has_table_names}
\alias{has_table_names}
\title{Factory of predicates to check if a table has the same names as a reference.}
\usage{
has_table_names(reference)
}
\arguments{
\item{reference}{A dataframe.}
}
\value{
A closure.
}
\description{
Factory of predicates to check if a table has the same names as a reference.
}
\examples{
stem <- data.frame(x = 1, y = 1)
tree <- data.frame(x = 1, z = 1)
has_table_names(stem)(stem)
has_table_names(stem)(tree)
}
\seealso{
Other predicates: \code{\link{check_crucial_names}},
  \code{\link{detect_if}}, \code{\link{is_multiple}}
}
\concept{predicates}
