% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insensitive.R
\name{insensitive}
\alias{insensitive}
\title{Create a vector or dataframe with case insensitive name matching.}
\usage{
insensitive(x)
}
\arguments{
\item{x}{vector or dataframe to modify.}
}
\description{
Create a vector or dataframe with case insensitive name matching.
}
\section{Acknowledgment}{

Thanks to Neil Richardson for recommending this function
(https://twitter.com/enpiar).
}

\examples{
x <- c("abc" = 1, "def" = 2)
x["ABC"]
y <- insensitive(x)
y["ABC"]
y[["ABC"]]
y$ABC

vft <- data.frame(TreeID = 1)
insensitive(vft)[["TreeID"]]
insensitive(vft)[["treeid"]]
insensitive(vft)["TreeID"]
insensitive(vft)["treeid"]
insensitive(vft)$TreeID
insensitive(vft)$treeid

# Works
transform(insensitive(vft), new = treeid)
# dplyr::mutate(insensitive(vft), new = treeid)
# Fails
# transform(insensitive(vft), new = TreeID)
# dplyr::mutate(insensitive(vft), new = TreeID)
}
\seealso{
\code{httr::insensitive()}.

Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}}, \code{\link{flag_if}},
  \code{\link{flag_multiple}}, \code{\link{multiple_var}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}, \code{\link{round_any}}

Other functions dealing with names.: \code{\link{name_dfs}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}, \code{\link{round_any}}
}
\author{
Hadley Wickham (see \code{httr::insensitive()}).
(https://twitter.com/hadleywickham)
}
\concept{functions dealing with names.}
\concept{functions for developers.}
