% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crucial_names.R
\name{is_named}
\alias{is_named}
\title{Is named object? (Base-only version of rlang::is_named).}
\usage{
is_named(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
The scalar \code{TRUE} or \code{FALSE}.
}
\description{
Is named object? (Base-only version of rlang::is_named).
}
