% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luquillo}
\alias{luquillo}
\alias{luquillo_stem_random_tiny}
\alias{luquillo_vft_4quad}
\title{Datasets from \strong{fgeo.data}.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 77 rows and 19 columns.}
\usage{
luquillo_stem_random_tiny

luquillo_vft_4quad
}
\description{
Datasets from \strong{fgeo.data}.
}
\examples{
str(luquillo_stem_random_tiny, give.attr = FALSE)
str(luquillo_vft_4quad, give.attr = FALSE)
}
\concept{ForestGEO datasets}
\keyword{datasets}
