% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_var.R
\name{multiple_var}
\alias{multiple_var}
\title{Factory of predicate functions to check for multiple values of a variable.}
\usage{
multiple_var(var)
}
\arguments{
\item{var}{Character string giving the name of a single variable.}
}
\value{
A single `TRUE` or `FALSE`. Insensitive to upper- lower-case.
}
\description{
Useful in `if()` statements.
}
\examples{
multiple_censusid <- multiple_var("censusid")
multiple_censusid(data.frame(CensusID = c(1, 2, NA)))
multiple_censusid(data.frame(CensusID = c(1, 1, NA)))

# Insensitive to upper/lowercase
multiple_censusid <- multiple_var("censusid")
multiple_censusid(data.frame(censusid = c(1, 2, NA)))
multiple_censusid <- multiple_var("CENSUSID")
multiple_censusid(data.frame(censusid = c(1, 2, NA)))
}
\seealso{
Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{flag_multiple}}, \code{\link{insensitive}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}, \code{\link{round_any}}

Other predicates.: \code{\link{extract_insensitive}}
}
\concept{functions for developers.}
\concept{predicates.}
\keyword{internal}
