% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_df_lst.R
\name{name_df_lst}
\alias{name_df_lst}
\title{Identify each dataframe in a list with a name.}
\usage{
name_df_lst(df_lst, name = "name")
}
\arguments{
\item{df_lst}{A list of dataframes.}

\item{name}{Names of the columns that store the names and values.}
}
\value{
A list of dataframes.
}
\description{
Identify each dataframe in a list with the name of the corresponding list
item.
}
\examples{
dfs <- list(a = data.frame(x = 1), b = data.frame(x = 1))
name_df_lst(dfs)

name_df_lst(dfs, "custom_name")

dfs2 <- list(data.frame(x = 1), data.frame(x = 1))
name_df_lst(dfs2)
}
