% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms_pull_matches.R
\name{nms_pull_matches}
\alias{nms_pull_matches}
\title{Pull names that match a character vector.}
\usage{
nms_pull_matches(x, .match)
}
\arguments{
\item{x}{A named object.}

\item{.match}{A character vector giving names to match.}
}
\value{
A character vector.
}
\description{
The \code{nms_} prefix matches functions in other fgeo packages.
}
\examples{
nms_pull_matches(luquillo_stem_random_tiny, c("x", "PX", "gx"))
nms_pull_matches(luquillo_vft_4quad, c("x", "PX", "gx"))
nms_pull_matches(luquillo_vft_4quad, c("PY", "PX", "gx", "gy"))
}
\seealso{
Other functions dealing with names.: \code{\link{insensitive}},
  \code{\link{name_dfs}}, \code{\link{rename_matches}},
  \code{\link{round_any}}

Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{detect_multiple_f}},
  \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{flag_duplicated}}, \code{\link{insensitive}},
  \code{\link{rename_matches}}, \code{\link{round_any}}
}
\concept{functions dealing with names.}
\concept{functions for developers.}
