% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms_restore_matching.R
\name{nms_restore_matching}
\alias{nms_restore_matching}
\title{Rename an object based on case-insensitive match of the names of a reference.}
\usage{
nms_restore_matching(new, old)
}
\arguments{
\item{new}{New object which names to restored if they match the reference.}

\item{old}{Named object to use as reference.}
}
\value{
The output is \code{new} with as many names changed as case-insensitive
matches there are with the reference.
}
\description{
Rename an object based on case-insensitive match of the names of a reference.
}
\examples{
ref <- data.frame(COL1 = 1, COL2 = 1)
new <- data.frame(col1 = 5, col2 = 1, n = 5)
nms_restore_matching(new, ref)
}
\seealso{
Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{check_unique}}, \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{insensitive}}, \code{\link{round_any}}
}
\concept{functions for developers.}
