% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%||\%}
\title{Default value for \code{NULL}.}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{If x is NULL, will return y; otherwise returns x.}
}
\description{
This infix function makes it easy to replace \code{NULL}s with a default value.
It's inspired by the way that Ruby's or operation (||) works. It is slightly
adapted from \code{rlang::}%||%`.
}
\examples{
1 \%||\% 2
NULL \%||\% 2
}
