% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_all.R
\name{print_all}
\alias{print_all}
\title{Print all rows and columns of a tibble -- as if it was a dataframe.}
\usage{
print_all(x, n = nrow(x), width = Inf)
}
\arguments{
\item{x}{A tibble}

\item{n}{Integer; number of rows to print.}

\item{width}{Integer; number of columns to print.}
}
\value{
A tibble.
}
\description{
Print all rows and columns of a tibble -- as if it was a dataframe.
}
\examples{
\dontrun{
if (!requireNamespace("tibble")) {
  stop(
    "To run this example you must first install tibble with:\\n",
    "install.packages('tibble')",
    call. = FALSE
  )
}
tbl <- tibble::tibble(runif(50))
print_all(tbl)
}
}
\seealso{
Other general miscellaneous functions: \code{\link{op-null-default}}
}
\concept{general miscellaneous functions}
