% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_name_matches.R
\name{pull_name_matches}
\alias{pull_name_matches}
\title{Pull names that match a character vector.}
\usage{
pull_name_matches(x, .match)
}
\arguments{
\item{x}{A named object.}

\item{.match}{A character vector giving names to match.}
}
\value{
A character vector.
}
\description{
Pull names that match a character vector.
}
\examples{
pull_name_matches(luquillo_stem_random_tiny, c("x", "PX", "gx"))
pull_name_matches(luquillo_vft_4quad, c("x", "PX", "gx"))
pull_name_matches(luquillo_vft_4quad, c("PY", "PX", "gx", "gy"))
}
\seealso{
Other functions dealing with names.: \code{\link{insensitive}},
  \code{\link{name_dfs}}, \code{\link{rename_matches}}

Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{check_unique}}, \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{insensitive}}, \code{\link{rename_matches}},
  \code{\link{round_any}}
}
\concept{functions dealing with names.}
\concept{functions for developers.}
