% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_plotdim.R
\name{regex_line}
\alias{regex_line}
\title{Add regex line-start and -end.}
\usage{
regex_line(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A character vector.
}
\description{
Add regex line-start and -end.
}
\examples{
regex_line(c("a", "b"))
}
