% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_plotdim.R
\name{round_any}
\alias{round_any}
\title{Round to multiple of any number. Copied from \code{plyr:::round_any.numeric()}.}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{Numeric vector to round.}

\item{accuracy}{Number to round to.}

\item{f}{Rounding function: floor, ceiling or round.}
}
\description{
Round to multiple of any number. Copied from \code{plyr:::round_any.numeric()}.
}
\examples{
# From pryr::round_any()
round_any(135, 10)
round_any(135, 100)
round_any(135, 25)
round_any(135, 10, floor)
round_any(135, 100, floor)
round_any(135, 25, floor)
round_any(135, 10, ceiling)
round_any(135, 100, ceiling)
round_any(135, 25, ceiling)
}
\seealso{
\code{plyr::round_any()} and \url{http://bit.ly/2JrBQK3}.

Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{insensitive}}, \code{\link{is_multiple}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}

Other functions dealing with names.: \code{\link{insensitive}},
  \code{\link{name_dfs}}, \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}
}
\concept{functions dealing with names.}
\concept{functions for developers with no dependencies.}
\concept{functions for developers.}
