% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_plotdim.R
\name{stick}
\alias{stick}
\alias{or}
\alias{commas}
\title{Paste wrappers.}
\usage{
or(...)

commas(...)
}
\arguments{
\item{...}{Arguments passed on to \code{base::paste}
\describe{
  \item{sep}{a character string to separate the terms.  Not
    \code{\link{NA_character_}}.}
  \item{collapse}{an optional character string to separate the results.  Not
    \code{\link{NA_character_}}.}
}}
}
\value{
String.
}
\description{
Paste wrappers.
}
\examples{
or(1:3)
}
