% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_manipulate_strings.R
\name{str_to_tidy_names}
\alias{str_to_tidy_names}
\title{Create a names-friendly version of a string (lowercase and with no spaces).}
\usage{
str_to_tidy_names(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A modified version of \code{x}.
}
\description{
Create a names-friendly version of a string (lowercase and with no spaces).
}
\examples{
messy <- "Hi yOu"

str_to_tidy_names(messy)

messy_named_string <- c(`Messy Name` = messy)
# Targets strings, not its names (Target names with `fgeo.tool::nms_tidy()`)
str_to_tidy_names(messy_named_string)

dfm <- data.frame(1)
setNames(dfm, str_to_tidy_names(messy))

# Makes more sense when operating on strings
setNames(list(1), str_to_tidy_names(messy))
}
\seealso{
\code{fgeo.tool::nms_tidy()}.
}
