% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{flag_if}
\alias{flag_if}
\title{Flag if a vector or dataframe column meets a condition.}
\usage{
flag_if(.x, .x_var, .if, .flag = warning, msg = NULL)
}
\arguments{
\item{.x_var}{String giving a column name of \code{.x}.}

\item{.if}{A condition on \code{x} or .x_var.}

\item{.flag}{A function that throws a condition e.g. \code{stop}, \code{warning},
\code{message}, \code{rlang::abort}, \code{rlang::warn}, \code{rlang::inform}.}

\item{msg}{An optional custom message.}

\item{x, .x}{A vector or dataframe.}
}
\value{
Throws a condition determined by \code{.flag} or invisible.
}
\description{
These funcitons are useful to create messagees, warnings and error messages
based on a condition.
}
\examples{
dupl <- c(1, 1)
# Flags
is_duplicated <- function(x) any(duplicated(x))
flag_vector_if(dupl, is_duplicated(dupl), message)
flag_vector_if(dupl, is_duplicated(dupl), message, "Duplicated values.")
# Silent
is_multiple <- function(x) length(unique(x)) > 1
flag_vector_if(dupl, is_multiple(dupl), message)
flag_vector_if(dupl, is_multiple(dupl), message, "Multiple values.")

mult <- c(1, 2)
# Silent
flag_vector_if(mult, is_duplicated(mult), message)
flag_vector_if(mult, is_duplicated(mult), message, "Duplicated values.")

# Flags
flag_vector_if(mult, is_multiple(mult), message)
flag_vector_if(mult, is_multiple(mult), message, "Multiple values.")

dupl <- c(1, 1)
dupl_df <- data.frame(dupl)
# Flags
is_duplicated <- function(x) any(duplicated(x))
flag_if(dupl_df, "dupl", is_duplicated(dupl), message)
flag_if(dupl_df, "dupl", is_duplicated(dupl), message, "Duplicated values.")
# Silent
is_multiple <- function(x) length(unique(x)) > 1
flag_if(dupl_df, "dupl", is_multiple(dupl), message)
flag_if(dupl_df, "dupl", is_multiple(dupl), message, "Multiple values.")

mult_df <- data.frame(mult)
# Silent
flag_if(mult_df, "mult", is_duplicated(mult), message)
flag_if(mult_df, "mult", is_duplicated(mult), message, "Duplicated values.")
# Flags
flag_if(mult_df, "mult", is_multiple(mult), message)
flag_if(mult_df, "mult", is_multiple(mult), message, "Multiple values.")
}
\seealso{
Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{flag_multiple}}, \code{\link{insensitive}},
  \code{\link{multiple_var}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}, \code{\link{round_any}}
}
\concept{functions for developers.}
