% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_multiple.R
\name{flag_if}
\alias{flag_if}
\title{Flag if a predicate returns TRUE and throw a condtion with optional message.}
\usage{
flag_if(.data, predicate, condition = warning, msg = NULL)
}
\arguments{
\item{.data}{Vector.}

\item{predicate}{A predicate function.}

\item{condition}{A condition function (e.g. \code{\link[=stop]{stop()}}, \code{\link[=warning]{warning()}},
\code{rlang::inform()}).}

\item{msg}{String. An optional custom message.}
}
\value{
A condition (and \code{.data} invisibly).
}
\description{
Flag if a predicate returns TRUE and throw a condtion with optional message.
}
