% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_multiple.R
\name{is_multiple}
\alias{is_multiple}
\alias{is_duplicated}
\alias{flag_duplicated}
\alias{flag_multiple}
\title{Detect and flag duplicated and multiple values of a variable.}
\usage{
is_multiple(.data)

is_duplicated(.data)

flag_duplicated(.data, cond, msg = NULL)

flag_multiple(.data, cond, msg = NULL)
}
\arguments{
\item{.data}{A vector.}

\item{cond}{Symbol; the bare name of a function that outputs a condition:
e.g. warning, stop, message, rlang::warn, rlang::abort, rlang::inform.}

\item{msg}{(Argument to the resulting function) String; an optional custom
message.}
}
\value{
See description.
}
\description{
\itemize{
\item \code{is_multiple()} and \code{is_duplicated()} return \code{TRUE} if they detect,
respectively, multiple different values of a variable (e.g. c(1, 2)\code{), or duplicated values of a variable (e.g. c(1, 1)}).
\item \code{flag_multiple()} and \code{flag_duplicated()} throw a condition (message,
warning, or error) when \code{is_multiple()} and \code{is_duplicated()} would
return \code{TRUE}. They also return the main input, invisibly.
}
}
\examples{
# DETECT -----------------------------------------------------------------
is_multiple(c(1, 2))
is_multiple(c(1, 1))
is_multiple(c(1, NA))

is_duplicated(c(1, 2))
is_duplicated(c(1, 1))
is_duplicated(c(1, NA))

# FLAG -------------------------------------------------------------------
duplicated_not_multiple <- c(1, 1, 1)
flag_multiple(duplicated_not_multiple, warning)

flag_duplicated(duplicated_not_multiple, warning)
flag_duplicated(duplicated_not_multiple, warning, "Custom message")
}
\seealso{
\code{\link[=detect_multiple_f]{detect_multiple_f()}}, \code{\link[=flag_multiple_f]{flag_multiple_f()}}, \code{\link[=detect_duplicated_f]{detect_duplicated_f()}},
\code{\link[=flag_duplicated_f]{flag_duplicated_f()}}.

Other functions to check inputs.: \code{\link{check_crucial_names}},
  \code{\link{detect_multiple_f}}

Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{detect_multiple_f}},
  \code{\link{exists_in_pkg}},
  \code{\link{extract_insensitive}},
  \code{\link{insensitive}},
  \code{\link{nms_pull_matches}},
  \code{\link{rename_matches}}, \code{\link{round_any}}

Other predicates.: \code{\link{detect_multiple_f}},
  \code{\link{extract_insensitive}}
}
\concept{functions for developers.}
\concept{functions to check inputs.}
\concept{predicates.}
