% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_plotdim.R
\name{nms_extract_match}
\alias{nms_extract_match}
\title{Extract names that match a character vector.}
\usage{
nms_extract_match(x, .match)
}
\arguments{
\item{x}{A named object.}

\item{.match}{A character vector giving names to match.}
}
\value{
A character vector.
}
\description{
Extract names that match a character vector.
}
\examples{
nms_extract_match(fgeo.data::luquillo_vft_4quad, c("x", "PX", "gx"))
nms_extract_match(fgeo.data::luquillo_stem6_1ha, c("x", "PX", "gx"))
}
